Replaced the update-time command sequence with update-list.
If the $(shell ...) function is invoked standalone (i.e. not as the right-
hand side of an assignment or in an ifeq construct), GNU make complains with
"*** unterminated call to function `shell': missing `)'.  Stop." if it
contains multiple commands separated by ";", although it is happy to
process a single command.